(chore) rename hook file query-client->useQueryClient
